Fix temporal layer log string.
BUG=
R=pthatcher@webrtc.org

Review URL: https://webrtc-codereview.appspot.com/43639004

Patch from Noah Richards <noahric@chromium.org>.

Cr-Commit-Position: refs/heads/master@{#8687}
git-svn-id: http://webrtc.googlecode.com/svn/trunk@8687 4adac7df-926f-26a2-2b94-8c16560cd09d

+++ talk/media/webrtc/simulcast.cc
@@ -401,7 +401,9 @@ void LogSimulcastSubstreams(const webrtc::VideoCodec& codec) {
                  << codec.simulcastStream[i].height << "@"
                  << codec.simulcastStream[i].minBitrate << "-"
                  << codec.simulcastStream[i].maxBitrate << "kbps"
-                 << " with " << codec.simulcastStream[i].numberOfTemporalLayers
+                 << " with "
+                 << static_cast<int>(
+                     codec.simulcastStream[i].numberOfTemporalLayers)
                  << " temporal layers";
   }
 }
